Rest API Project files

+++ Python Programming/RestAPIs/resources/item.py
@@ -1,6 +1,5 @@
-import uuid
-from flask import request
 from flask.views import MethodView
+from flask_jwt_extended import jwt_required, get_jwt
 from flask_smorest import Blueprint, abort
 
 from db import db
@@ -13,14 +12,20 @@ from sqlalchemy.exc import SQLAlchemyError
 blp = Blueprint("Items", __name__, description="Operations on items")
 
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
+
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
          item = ItemModel.query.get_or_404(item_id)
          return item
-
+    
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -44,11 +49,13 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
 
-   
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
